+++ src/core/plugin/enums/plugin-selector-node-type.enum.ts
@@ -1,0 +1,14 @@
+export enum PluginSelectorNodeType {
+	Tag = 'tag',
+	String = 'string',
+	Selector = 'selector',
+	Root = 'root',
+	Pseudo = 'pseudo',
+	Nested = 'nested',
+	ID = 'id',
+	Comment = 'comment',
+	Combinator = 'combinator',
+	Class = 'class',
+	Attribute = 'attribute',
+	Universal = 'universal',
+}