+++ src/core/plugin/plugins/plugin-selector-combinator-nesting/interfaces/plugin-selector-combinator-nesting.interface.ts
@@ -1,0 +1,2 @@
+export type PluginSelectorCombinatorNestingName = string;
+export type PluginSelectorCombinatorNestingMessageArgs = [PluginSelectorCombinatorNestingName];