+++ src/core/plugin/plugins/plugin-selector-combinator-nesting/providers/plugin-selector-combinator-nesting.provider.spec.ts
@@ -1,0 +1,18 @@
+import { PluginConfigExecutionMode } from '../../../interfaces/plugin-config.interface';
+
+import { stylelintMock } from '../../../../mocks/stylelint.mock';
+import { PluginSelectorCombinatorNesting } from '../api/plugin-selector-combinator-nesting';
+import { pluginSelectorCombinatorNestingProvider } from './plugin-selector-combinator-nesting.provider';
+
+jest.mock('stylelint', () => stylelintMock());
+
+describe('PluginSelectorCombinatorNestingProvider', () => {
+	it('should have the following value', () => {
+		const provider = {
+			provide: PluginSelectorCombinatorNesting,
+			options: 'always' as PluginConfigExecutionMode,
+		};
+
+		expect(pluginSelectorCombinatorNestingProvider()).toEqual(provider);
+	});
+});