+++ src/index.ts
@@ -80,6 +80,7 @@ class Configuration implements Config {
 		'selector-no-qualifying-type': true,
 		'selector-not-notation': 'simple',
 		'selector-pseudo-element-no-unknown': [true, { ignorePseudoElements: ['ng-deep'] }],
+		'selector-nested-pattern': '^(?:&:?[^&]+|[^&:]+)$',
 
 		/* Time */
 		'time-min-milliseconds': 50,